change datatype of id in Model to Integer

+++ src/com/polymart/model/AnhSanPhamModel.java
@@ -1,22 +1,22 @@
 package com.polymart.model;
 
 public class AnhSanPhamModel {
-    private Long idSanPham;
+    private Integer idSanPham;
     private String tenAnh;
 
     public AnhSanPhamModel() {
     }
 
-    public AnhSanPhamModel(Long idSanPham, String tenAnh) {
+    public AnhSanPhamModel(Integer idSanPham, String tenAnh) {
         this.idSanPham = idSanPham;
         this.tenAnh = tenAnh;
     }
 
-    public Long getIdSanPham() {
+    public Integer getIdSanPham() {
         return idSanPham;
     }
 
-    public void setIdSanPham(Long idSanPham) {
+    public void setIdSanPham(Integer idSanPham) {
         this.idSanPham = idSanPham;
     }
